refactor instrumented test utils location
