+++ app/src/androidTest/java/com/google/jetpackcamera/utils/AppTestUtil.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera
+package com.google.jetpackcamera.utils
 
 import android.os.Build
